# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent-Native Middleware API",
    "description": "## Agent-Native Middleware API\n\nHeadless infrastructure for the B2A (Business-to-Agent) economy. This API provides four core services that give AI agents the 'hands' and 'eyes' to manipulate the digital and physical world.\n\n### Service Pillars\n\n- **IoT Protocol Bridge** (`/v1/iot`) \u2014 Secure, ACL-enforced protocol translation for IoT devices (MQTT, CoAP, Zigbee, etc.)\n- **Autonomous Product Manager** (`/v1/telemetry`) \u2014 Ingest telemetry, detect anomalies, auto-generate pull requests to fix bugs\n- **Programmatic Media Engine** (`/v1/media`) \u2014 Video ingestion, viral hook detection, reframing, captioning, and cross-platform distribution\n- **Agent Communications** (`/v1/comms`) \u2014 Agent registration, structured messaging, capability discovery, and swarm task handoffs\n\n### Authentication\n\nAll endpoints require an API key passed via the `X-API-Key` header.\n\n### For Agents\n\nThis API is designed for programmatic consumption. See `/llm.txt` for LLM-optimized documentation and `/openapi.json` for the full spec.",
    "contact": {
      "name": "Agent-Native Middleware",
      "email": "api@yourdomain.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.yourdomain.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
...
    }
  ],
  "paths": {
...
    "/v1/telemetry/events": {
      "post": {
        "tags": [
          "Autonomous Product Manager"
        ],
        "summary": "Ingest telemetry events",
        "description": "Submit a batch of telemetry events for anomaly detection. Events are processed asynchronously. The Autonomous PM will analyze patterns across events to identify error spikes, latency regressions, and missing feature patterns.",
        "operationId": "ingest_events_v1_telemetry_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryBatchResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/telemetry/events/single": {
      "post": {
        "tags": [
          "Autonomous Product Manager"
        ],
        "summary": "Ingest a single telemetry event",
        "description": "Convenience endpoint for submitting one event at a time.",
        "operationId": "ingest_single_event_v1_telemetry_events_single_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/telemetry/anomalies": {
      "get": {
        "tags": [
          "Autonomous Product Manager"
        ],
        "summary": "List detected anomalies",
        "description": "Retrieve anomalies detected by the Autonomous PM. Anomalies are classified by severity and category, with optional LLM-generated fix suggestions and auto-PR links.",
        "operationId": "list_anomalies_v1_telemetry_anomalies_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/app__schemas__telemetry__Severity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity level",
              "title": "Severity"
            },
            "description": "Filter by severity level"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/telemetry/anomalies/{anomaly_id}": {
      "get": {
        "tags": [
          "Autonomous Product Manager"
        ],
        "summary": "Get anomaly details",
        "description": "Retrieve detailed information about a specific anomaly, including suggested fixes.",
        "operationId": "get_anomaly_v1_telemetry_anomalies__anomaly_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "anomaly_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Anomaly Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyReport"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/telemetry/anomalies/{anomaly_id}/auto-pr": {
      "post": {
        "tags": [
          "Autonomous Product Manager"
        ],
        "summary": "Generate an autonomous pull request",
        "description": "Instruct the Autonomous PM to generate a code fix for the given anomaly and optionally push it as a pull request. Use dry_run=true to preview the proposed diff without committing.",
        "operationId": "generate_auto_pr_v1_telemetry_anomalies__anomaly_id__auto_pr_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "anomaly_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Anomaly Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoPRRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoPRResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/telemetry/stats": {
      "get": {
        "tags": [
          "Autonomous Product Manager"
        ],
        "summary": "Get telemetry statistics",
        "description": "Aggregate statistics across all ingested telemetry events. Useful for agents monitoring system health at a glance.",
        "operationId": "get_stats_v1_telemetry_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnomalyListResponse": {
        "properties": {
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyReport"
            },
            "type": "array",
            "title": "Anomalies"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "anomalies",
          "total",
          "page",
          "per_page"
        ],
        "title": "AnomalyListResponse",
        "description": "Paginated anomaly listing."
      },
      "AnomalyReport": {
        "properties": {
          "anomaly_id": {
            "type": "string",
            "title": "Anomaly Id"
          },
          "severity": {
            "$ref": "#/components/schemas/app__schemas__telemetry__Severity"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Classification of the anomaly (e.g., 'error_spike', 'latency_regression', 'missing_feature')."
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Agent-readable summary of the issue."
          },
          "affected_endpoints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Endpoints",
            "description": "API endpoints or services affected."
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "description": "Number of telemetry events contributing to this anomaly."
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "suggested_fix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Fix",
            "description": "LLM-generated code fix suggestion in diff format."
          },
          "auto_pr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Pr Url",
            "description": "URL of the auto-generated pull request, if AUTO_PR_ENABLED."
          }
        },
        "type": "object",
        "required": [
          "anomaly_id",
          "severity",
          "category",
          "summary",
          "event_count",
          "first_seen",
          "last_seen"
        ],
        "title": "AnomalyReport",
        "description": "An anomaly detected by the Autonomous PM from telemetry analysis."
      },
...
      "AutoPRRequest": {
        "properties": {
          "anomaly_id": {
            "type": "string",
            "title": "Anomaly Id",
            "description": "The anomaly to generate a fix for."
          },
          "target_repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Repo",
            "description": "Git remote URL. Uses system default if omitted."
          },
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name",
            "description": "Branch name for the PR. Auto-generated if omitted."
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "If true, returns the proposed diff without pushing.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "anomaly_id"
        ],
        "title": "AutoPRRequest",
        "description": "Request the Autonomous PM to generate and push a fix."
      },
      "AutoPRResponse": {
        "properties": {
          "anomaly_id": {
            "type": "string",
            "title": "Anomaly Id"
          },
          "pr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Url",
            "description": "URL of the created PR (null if dry_run=true)."
          },
          "diff": {
            "type": "string",
            "title": "Diff",
            "description": "The proposed code changes in unified diff format."
          },
          "files_changed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files Changed"
          },
          "tests_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests Passed",
            "description": "Whether auto-generated unit tests passed."
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "anomaly_id",
          "diff",
          "files_changed",
          "status"
        ],
        "title": "AutoPRResponse",
        "description": "Result of an autonomous pull request generation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelemetryBatch": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/TelemetryEvent"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Events",
            "description": "Array of telemetry events. Max 100 per batch. RED TEAM FIX: Reduced from 1000 to prevent memory exhaustion attacks. For higher throughput, send multiple batches."
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "TelemetryBatch",
        "description": "Batch submission of telemetry events for efficient ingestion."
      },
      "TelemetryBatchResponse": {
        "properties": {
          "ingested": {
            "type": "integer",
            "title": "Ingested"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors",
            "description": "Details of any events that failed validation."
          }
        },
        "type": "object",
        "required": [
          "ingested",
          "failed",
          "batch_id"
        ],
        "title": "TelemetryBatchResponse",
        "description": "Confirmation of batch ingestion."
      },
      "TelemetryEvent": {
        "properties": {
          "event_type": {
            "$ref": "#/components/schemas/TelemetryEventType"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Originating service or module name.",
            "examples": [
              "iot-bridge",
              "media-engine",
              "auth-service"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human/agent-readable description of the event."
          },
          "severity": {
            "$ref": "#/components/schemas/app__schemas__telemetry__Severity",
            "default": "info"
          },
          "stack_trace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Trace",
            "description": "Full stack trace for error events."
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Structured context (request_id, user_agent, latency_ms, etc.)."
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "Event timestamp. Server will assign one if omitted."
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "source",
          "message"
        ],
        "title": "TelemetryEvent",
        "description": "A single telemetry event to ingest."
      },
      "TelemetryEventType": {
        "type": "string",
        "enum": [
          "error",
          "warning",
          "session",
          "api_call",
          "llm_trace",
          "performance",
          "custom"
        ],
        "title": "TelemetryEventType",
        "description": "Categories of telemetry events the system can ingest."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__telemetry__Severity": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low",
          "info"
        ],
        "title": "Severity",
        "description": "Severity levels for anomalies and issues."
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key for agent authentication. Pass in the X-API-Key header.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}